fix(SCRUM-18): add login wait for home page

+++ src/test/java/Hooks/Hook.java
@@ -3,29 +3,43 @@ package Hooks;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
 
 public class Hook {
 
     public static WebDriver driver;
+
     @Before
-    public void setup(){
+    public void setup() {
         ChromeOptions options = new ChromeOptions();
         options.addArguments("--remote-allow-origins=*");
         options.addArguments("--incognito");
         options.addArguments("--ignore-certificate-errors");
+
         WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+        driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(60));
+        driver.manage().timeouts().scriptTimeout(Duration.ofSeconds(30));
+
         driver.get("https://automationexercise.com/");
         driver.manage().window().maximize();
+
+        new WebDriverWait(driver, Duration.ofSeconds(20)).until(webDriver ->
+                "complete".equals(((JavascriptExecutor) webDriver).executeScript("return document.readyState")));
     }
 
     @After
-    public void quit(){
-        driver.manage().deleteAllCookies();
-        driver.quit();
-
+    public void quit() {
+        if (driver != null) {
+            driver.manage().deleteAllCookies();
+            driver.quit();
+        }
     }
 }
